refactor dropdown
