+++ optimus/lib/src/dropdown/common.dart
@@ -1,0 +1,1 @@
+const itemMinHeight = 69.0;
